[IMP] social: when login a social account, set the right company
Purpose
=======
We want to set the right company on the social account during the
authentication.

Technical
=========
We can not use the context or "self.env.company" because those
information are lost when the user is redirected to the social media
website (Facxebook, Twitter...) and then redirected again to Odoo.

So we store this information in the user session, which will still be
the same even after the redirections.

Task 2415987
See odoo/enterprise/pull/15896
See odoo/upgrade/pull/2099