[IMP] test_l10n_be_hr_payroll_account: change the name in hr_contract_salary_tour_2
In the previous test, the name "Mitchell Admin 2" is used in hr_contract_salary_tour_2.
When "Create Employee" is clicked, a new partner named "Mitchell Admin 2" will be created.
However, a partner with the same name has already been created in the previous test hr_contract_salary_tour.

In order to avoid confusion in the following test (e.g. search a partner by its name),
the name is changed to "Mitchell Admin 3"