[IMP] crm_enterprise: improve/re-arrange KPIs on dashboard view
Right now, on the dashboard view of leads, the KPIs are not in
optimum order.

This commit improves the behaviour by putting the related KPIs
together (based on whether the leads are activated or not) and
thus improves the reporting.

Apart from that, this commit also adds MRR related KPIs if
recurring revenue is activated.

updated aggregate name, purpose is to have a global convention
and avoid name clashing, updated aggregate names are listed below.

- aggregate_leads
- aggregate_opportunities
- aggregate_days_to_convert
- aggregate_days_to_assign
- aggregate_expected_revenue
- aggregate_expected_monthly_revenue
- aggregate_prorated_revenue
- aggregate_prorated_monthly_revenue
- aggregate_day_close
- aggregate_days_exceeding_closing

TaskId-2343223

Part-of: odoo/enterprise#18797