[FIX] quality,mrp_workorder: avoid no-scalable domain
Making a dynamic domain with a `allowed_xxx_ids` field is a bad idea
when the domain doesn't filter out the most of records. It can create a
performance issue, if the `allowed_xxx_ids` contains too much ids
(for 70K by example), then a lot of data will be exchange to/from server
for each onchange (500Ko, > 2 sec in that case) which slows down all
the form views (and the JS take a lot of time to parse/use the field).

Change it into a static domain, the onchange become instantaneous and
the name_search is 4 times faster.

Also clean/simplify the `_compute_component_ids`.

task-2558097

closes odoo/enterprise#18849

Related: odoo/upgrade#2538
Related: odoo/odoo#71870
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>