[FIX] documents_spreadsheet: Select Field matching before related model
Currently there is a traceback when adding a filter, selecting a field
matching and then selecting a related model. From a functional point of vue it makes
no sense to choose a field match before the related model. Therefore this
commit will hide the field matching part when no related model is selected.

closes odoo/enterprise#20381

Task-id: 2362089
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>