[FIX] account_asset: create model from asset
An asset model created from an asset is not usable.

To reproduce the error:
1. Create an asset
2. Confirm the asset
3. Save Model
4. Create a new asset:
    Error: the model is not available
5. Accounting > Configuration > Asset Models
    Error: the model is not present

When consulting an asset, the key `default_asset_type` and `asset_type`
are in the context. However, when clicking the action "Save Model",
several keys are added:
https://github.com/odoo/enterprise/blob/b802b4b856a7ca74f390b78d9cd9fdf52e81b6a5/account_asset/models/account_asset.py#L482-L502
And the old context keys are filtered:
https://github.com/odoo/odoo/blob/44b59d10d0ab6440b0a908f518c157a3d670627b/addons/web/static/src/js/chrome/action_manager_act_window.js#L661-L670
Therefore `default_asset_type` is no more available.

Later, when saving the asset model, since the default value of
`asset_type` is not in the context, the field of the new record is not
defined. Then, the method `flush` triggers the methods `compute`, but
without all context keys mentioned above (`asset_type` is not
available). The field will thus stay undefined:
https://github.com/odoo/enterprise/blob/b802b4b856a7ca74f390b78d9cd9fdf52e81b6a5/account_asset/models/account_asset.py#L187-L195
This is the reason why the model won't be usable.

OPW-2467962

closes odoo/enterprise#18888

X-original-commit: aad3d0ea285a9bf6df4c6bdf68293437f2f3dc93
Signed-off-by: William André (wan) <wan@odoo.com>