[FIX] l10n_es_reports: speedup computation of `l10n_es_reports_mod349_available` on install
On installation of the module `l10n_es_reports`,
the field `l10n_es_reports_mod349_available` gets computed
for all existing invoices, and if there are thousands of them
it can take a while.

In the case of the upgrade request #8578,
it was taking more than 72 hours for all invoices,
and counting.

Using raw SQL, it's a matter of seconds to set
this computed field.

upg-8578

closes odoo/enterprise#18831

X-original-commit: 23b5f550b539868c6bb898ab99793790243c0f5d
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>