[FIX] account_reports: fix journals filter when opening journal items
- Go to Accounting > Reporting > Trial Balance
- Select 2 journals in Journals filter (i.e. Bank and Customer Invoices)
- Click on an account linked to the second journal (i.e. 101200 Account Receivable)
  and select "Journal Items"
An empty list of journal items will be displayed.

When opening journal items, "search_default_account_id" is set to selected account and
"search_default_journal_id" is set to selected journals.
But "search_default_" with a many2one field doesn't support multiple values and in case
of an array, it will take the first element.
In our case, journal items are filtered by the selected account (i.e Account Receivable)
and the first element of the selected journals (i.e. Bank), which is an undesired
combination.

In case of several journals selected, domain will be used instead of "search_default_journal_id".

opw-2507576

closes odoo/enterprise#18951

X-original-commit: fa28dee19892f5ac41f7e01eca66739dfc0cfcb3
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>