[FIX] stock_barcode: avoid fetching same barcode data twice
get_all_products_by_barcode may take significant time in large databases (10-20
seconds). No need to load same data if we already did it before.

It assumes that fetching result is the same as in previous request.

1. It may be different if the data are changed, e.g. product\location are added.
It should not happen often and can be sorted by refreshing page. It should be a
good trade off for performance

2. In multi-company environment the fetched data may be different.

2.1 In Odoo 12.0
it's possible to change current company in another browser tab and hence get
inconsistency. But again, it can be sorted by page refreshing.

2.2. In later Odoo versions to change company you are forced to refresh the page
and hence the cache will be cleaned.

---

opw-2355449

closes odoo/enterprise#19749

X-original-commit: 3d3c352310b46539c5a167b587f809891de012a2
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>
Signed-off-by: Ivan Yelizariev // IEL <yelizariev@users.noreply.github.com>