[FIX] point_of_sale: correct compute prices of order in OrderManagementScreen
When viewing an order in the order management screen that has customer
with fiscal position, the calculation of prices is based on the fiscal
position of the currently active order. This commit fixes the issue
by properly specifying the order in the calculation of taxes based
on the order's customer and/or fiscal position.

closes odoo/enterprise#19509

X-original-commit: dad7fde1a07e7997f43086568a2d7864a7af5872
Related: odoo/odoo#73396
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>