[FIX] account_reports: tax report: make "save" button work again
Before saas-14.4, trying to get id field of account.generic tax report returned False, and was converted to 0 when passing it as the value of the export wizard. Now, it raises an error, as this report does not have an id field (since it's an abstract model). We restore the intended behavior by simply passing 0 if the report doesn't have an id.

closes odoo/enterprise#18946

X-original-commit: d7d3e052ecfe8e46ba9a7e72aac51d4df875b7d5
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>