[FIX] web_enterprise: menus supports image as icon
With studio, edit an app icon and select an image.
Before this commit, there was a crash in python (load_menus).
It was because when done this way the menu's web_icon field was empty
(it contains css values for colors and pictogram when the icon is not
an image).
After this commit, load_menus route supports loading menus that have
an image as icon.