[FIX] project_timesheet_synchro: make tours wait for sync to finish
When the timesheet app starts, it tries to sync with the backend. The
application however does not wait for this sync to complete before
starting. As such, tours that run in it can start immediately.

The sync operation needs to make two successive RPCs, both of which need
the user to be authenticated. Because the tours are short, it can
sometimes happen that the tour will complete before the second RPC is
sent. When the test runner see that the tour is successful, it will
clear the cookies, including the session cookies, which will cause the
second RPC to crash because it doesn't have a valid session cookie
attached, which in turn will mark the test as failed.

This commit fixes that by adding an extra step in the tours (or
modifying the last step if it didn't have an extra trigger) where we
wait for the sync indicator to stop being displayed, indicating that the
sync is finished