[FIX] account_bank_statement_import_camt: several tx details
Suppose a CAMT file with one entry (`<Ntry>`) that contains some details
(`<NtryDtls>`). Suppose that, in details, there are several transactions
details (`<TxDtls>`). When importing the file, the module detects the
different transaction. However, when searching the amount for each
transaction, it will first find the amount of the entry. As soon as it
gets an amount, the search is stopped. This is the reason why the amount
used is incorrect.

When getting an amount, this fix suggests to inverse the search order:
always search from the most specific to the least specific.

OPW-2512363

closes odoo/enterprise#19080

X-original-commit: e95d537fd8be61411eeda4f5a0786ae704ea21e5
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>