[FIX] industry_fsm_sale: remove domain updated for project_template_id
Before this commit, the user cannot select a project with a pricing_type
different to task rate. The reason of this it is because we give the
same domain than the one in the project_id field.

This commit removes the updated domain for project_template_id in
product.product model in industry_fsm_sale module. Then, the user will
be able to select project with pricing_type different to task rate as it
was the case before.

X-original-commit: 1e088641f4555ff3455f55499eeed3a5ea62483b

Step to reproduce:
-----------------
1) create basic project A with a SOL set.
2) Create a service product with service_tracking='task_in_project'
3) Try select the project A in project_template_id.

Actual behaviour:
----------------
The project A is not shown because it is not a fsm project or project
with pricing_type='task_rate'.

Expected behaviour:
------------------
The project A should be shown in project_template_id because we want the
basic project with pricing type different to task rate. That is a
project with SOL and/or an employee mapping.

closes odoo/enterprise#18984

X-original-commit: 2262b93aa92500fd71b40b8c8c7ffb2289ced09f
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>