[FIX] document_spreadsheet: do not store filter model ID
Currently, we save globalFilters with the ID of model used to filter
(i.e. the id of the corresponding `ir.model` record) and used it when
loading the filter data inside the side panel.

Unfortunately, this is only relevant at runtime as the id of a given
model can change on every database, depending on the number of existing
ir.model and even their creation order.
This means that a spreadsheet cannot be shared accross similar databases
and might even break inside the same database in case of a backup/restore.

As we also stored the name of the model, we can take advantage of it
when loading the filter data and use the name to recover the model id.

Practically, this was breaking our demo templates that contained
filters as it is equivalent to loading a spreadsheet from another database.

closes odoo/enterprise#18977

Task: [2520389](https://www.odoo.com/web#id=2520389&action=333&active_id=2328&model=project.task&view_type=form&cids=1&menu_id=4720)
X-original-commit: 31a9d408c624e5977b54e37d8e60b1478700b790
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>