[IMP] sign: different access_token when cancel sign requests
Create different access_token for records in
	sign.request:
		action_draft() and action_cancel()
	sign.request.item:
		action_draft()

Before this task, the above functions will create the same access_token for all records in self.
There was no error, because all these functions are used only for a single record.
But these functions should support multiple records, so the problem should be fixed.

closes odoo/enterprise#17977

Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>