[FIX] web_grid: add padding in specific header of grid view
Before this commit, a CSS rule adds '/ ' when we have many fields in the
group by for the grid view. The problem is the selector CSS of this
rule. When the rule selects the following element in this selector:
`.o_view_grid > tbody > th > div > div:before`, we don't check the div
with a class and thus, when we use the `Many2OneAvatar` widget in the
timesheet grid, the rule is satisfied and add padding but the div of the
avatar is displayed as a block and thus, the avatar div go down in next
line because of this added padding.

This commit adds a class to the div concerned for the CSS rule, to only
apply the CSS rule when needed.

Step to see the additional padding before this commit:
=====================================================
1) Go to Timesheets App
2) Go to grid view in `Timesheet > All Timesheets` menu

Actual behaviour:
================
Above each avatar, we have an added padding.

Expected behaviour:
==================
The row containing the avatar should not have an additional padding, all
rows in the `tbody` should keep the same height.

closes odoo/enterprise#19051

X-original-commit: 3eb7697fd0bc41ae991d613e48f38fbb552d59f6
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>