[FIX] web_enterprise: open dialog on the home_menu
When a dialog is opened from the home_menu, the keydown events are
caught by the home_menu. This causes strange behavior
(loss of focus, etc) and prevents the hotkeys to work properly.

To solve this problem, the home_menu will not catch keydown events
if the main screen is not the activeElement.

closes odoo-dev/enterprise#183

Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>