[IMP] {,sale_}planning: add role on product
This commit adds the planning_enable and the planning_role properties to
product templates.

Those properties allows one to create a sale order with such products
and take advantage of some functionnalities in the gantt :
- Plan automatically the sale order using those product;
- Take into account product_qty while copying slots from week to week;
- etc.

With this commit, we flag products that should generate a shift when
confirming the SO and set a role.
The views are adapted in order to display the products configured with
one role.
Tests are added in order to test product form and configuration ability.
Demo data are added in order to have some examples.

PR community : odoo/odoo#75553
PR enterprise : #19647
PR upgrade : odoo/upgrade#2777

task-2559061 - part:1