[FIX] web_enterprise: form view statusbar's dropdown state on mobile
On mobile/smaller screens, in the Form view, once there is more than one
button in the statusbar, they are grouped into a dropdown to prevent
from cluttering/breaking the layout due to the restricted available
space.

In the first implementation of this feature, those buttons were merged
into a dropdown no matter how many buttons were present. The drawback
was that if only one button was present, the dropdown introduced an
unnecessary click/touch in the flow. A second iteration of this
implementation introduced the current conditional grouping... but also
added some complexity as every modifier (ie. invisible) could change de
number of buttons displayed, and in consequence, the need to group them
into a dropdown.

We discovered during the execution of the main flow tour on powerful
enough hardware (like the runbot) that some timing conjunction between
the actions performed in the UI and the responses of the server could
lead to an unexpected state of the dropdown.

More precisely:
- When the main flow tour mobile reaches the Vendor Bill creation step
- The newly created bill is edited to set the invoice_date
- This edit triggers both an 'onchange' and recompute the 'modifiers'
- Every application of the modifiers rerender the dropdown (if needed)
- But the 'onchange' response can also influence the modifiers
- If the dropdown is opened before the 'onchange' response, it will be
closed again once the onchange's triggers a rerender of the dropdown.
This case is less likely to happen due to a 'real user' interaction time
but may occur via the automation of the tours.

This commit fixes it by keeping the state (open/close) of the dropdown
at the FormRenderer's level to avoid this kind of timing issue and
prevent dropdown's flickering.

closes odoo/enterprise#19188

X-original-commit: 9b7f26435f3b193dcdb352a50b6a6ece9c781a8a
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>