[IMP] account_sepa: move sepa pain version from settings to journals
The goal is to move the SEPA pain version from the settings to each bank
account journal, and to make it editable by the user.

It will first look at IBAN, then at the fiscal country and lastly at the
company's country code to compute the initial value.

closes odoo/enterprise#19010

Task: 2567701
Related: odoo/upgrade#2562
Signed-off-by: William André (wan) <wan@odoo.com>