[FIX] web_enterprise: show correct message in expiration panel
The expiration panel needs to know if there is at least an app
installed. We rely on the 'mail' addon for this, as almost every
other addons depends on it. Before this commit, the JS code used
the 'module_list' key in session_info. However, this key has been
removed by [1].

This commit re-applies [2] in the master-wowl branch.

[1] odoo/odoo@7cfb0e0e748c9f02ecd43faf56dc8e5c3aca5e6d
[2] odoo/enterprise@d26ae4cb581dea06c6548a64aece1116d0bae0d6

closes odoo-dev/enterprise#184

Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>