[FIX] l10n_pe_edi: Error management on read_cdr is now properly parsed.
In the fix of avoid errors comming from SUNAT the error management was broken, now the
original message in the CDR is correctly shown, giving useful information o the user to fix
their documents.

Included a better help for debugging due to the dual way of receiving this information the reading process
of a CDR is a little bit complex, having the use cases in the documentation of the code will
be useful for maintainance.

closes odoo/enterprise#19988

X-original-commit: 9642873257e0475d1a5c470be3d259a7002399b5
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>