[FIX] l10n_pe_edi: Detraction amount must always be in PEN
The amount is already being calculated from the amount signed field
(which is always in PEN) but the currency in the template was inconsistently
taken from the document.

We only pass the banco de la nacion number if the bank and the bank account
are defined.  It will give an error later on from the government if we could
not find it.

closes odoo/enterprise#19446

X-original-commit: 277eed47ba7e3e41a164b313989da5af22aeb8dc
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Josse Colpaert <jco@openerp.com>