[REF] documents_spreadsheet: Split pivot side panel component
The pivot side panel component manages two things:
1) the list of pivots
2) the details of each pivot (domain, groupby, fields, etc.)

This is a preliminary work to to allow editing the domain.
This feature will add code and logic to the side panel component and we don't
want to drift towards a "giant pivot side panel component to rule them all".

This commit splits the current component in two components, each with a better
defined scope of responsabilities:
- one to manage the list of pivots and which pivot details to show
- one to actually manage pivot details

Task 2296880

closes odoo/enterprise#19019

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>