[IMP] sale_planning: link sol and planning shifts
With this commit, a relation is added between sol and shifts in order to
take advantage of the product configuration of the product set on the
sol.

A sol with a plannable product has a number of hours to plan and a
number of hours already planned. This is computed based on the slots
related to the SOL.

The slots related to a sol have some different behaviors than the
basic slots. They are unschedulable (this means they may have start
datetime and end datetime unset), and plannable (this means they may be
automatically scheduled regarding the sol product configuration)

At each moment there can't be more than one unscheduled slot for a given
sale order line. This must be checked after every write that can affect
planning slots. Also, the unscheduled planning slot, if it exists,
always have to be updated in order to show the hours remaining on the
sol.

In this commit, this new relation and the consequence that a
start_datetime and end_datetime can now be Falsy are handled.
Views are adapted to show this relation
Portal views are adapted to show this relation.

PR community : odoo/odoo#75553
PR enterprise : #19647
PR upgrade : odoo/upgrade#2777

task-2559061 - part:2