[FIX] mrp_workorder: label color contrast
Since https://github.com/odoo-dev/odoo/commit/845dc535871a9454317b0b8ebbe9de4eecb18bf7 ,
the color of the form_label was forced to the heading_color (black in
this case), and the contrast on the mrp workorder wasn't big enough.

Now, on the mrp workorder the form_lable is forced to the
o-wo-tablet-text color which will increase the contrast.

closes odoo-dev/enterprise#186

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>