[IMP] sale_subscription: no longer keep draft invoices for pending txs
Before this commit, a draft invoice was created to compute the amount
and currency of the payment transaction when renewing a subscription.
The invoice was posted as soon as the transaction was confirmed and
deleted if the transaction ended in a final blocking state. If the
transaction reached the pending state, the invoice was indefinitely kept
on hand in case the transaction was eventually confirmed. This
eventuality could not be ruled out because it might have happened that
the customer had to go through a 3DS authentication flow or that the
payment was not immediately confirmed.

Since commit 139dd9d, it is possible to renew a subscription by paying
with a 'redirection' flow. Because this flow increases the friction of
the payment process, customers are more likely to quit the flow which
means that the transaction will remain in pending state forever. The
combination of this change with the progressive adoption of the new PSD2
regulation (less chance of confirming the payment) leads to an increase
of draft invoices never being posted, hence polluting the database.

This commit makes it so that invoices created to compute the amount and
currency are immediately deleted after they fulfilled their role. A new
invoice is created on the fly and immediately posted only when the
payment is confirmed.

task-2494916

closes odoo/enterprise#19074

X-original-commit: aacfdba092b9f48e3047506e39db5ffc006258c8
Signed-off-by: Antoine Vandevenne (anv) <AntoineVDV@users.noreply.github.com>