[IMP] sign: refuse to sign
add allow refuse check in the sign.send.request
add refuse button in the signing page, a user can see it when
show a refuse confirm and add a textarea for the refuse reason
add refuse reason in chatter log
add refuse in sign.log
send mail to all signers with the refuse reason
when a sign.request is refused by a signer, other signers cannot sign it anymore

When someone is asked to sign a document that is not correct, his only solution is to let the signature request pending and contact the sender.
it's a loss of time for the people that are asked to sign, because they don't know that the document doesn't meet the requirements, or is wrong.
In other software, you've got a functionnality called "refuse to sign".
We need to offer that possibility to the users of Sign.

closes odoo/enterprise#19972

Taskid: 2614105
Related: odoo/upgrade#2704
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>