[FIX] helpdesk_sale_timesheet: fix customer selection issue
Currently, when go to helpdesk > configuration > helpdesk teams >
enable 'timesheets on tickets' and 'time reinvoicing'
then create a ticket > set a customer > save > remove the selected
the customer then the customer is set again and can not be unselected due
to recursive compute of sale_order_id and commercial_partner_id on
ticket.

So in this commit, set the so false when SO does not map with the SO of
So line/project/task so it will not set the partner as recursively.

closes odoo/enterprise#19391

Taskid: 2535773
X-original-commit: fdf438ed838e78545a6fdc3044907b6f1225188a
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>