[FIX] l10n_mx_edi_landing: check the custom number only if it's filled
Steps to reproduce the bug:
- Install the l10n_mx_edi_landing module
- Go to inventory > Operations > Landed Costs
- Create a new landed cost > add product
- Do not fill the “customs number” field
- An error is triggered

Problem:
A constraint is checked for each landed cost, in which we try to directly access the 'l10n_mx_edi_customs_number' field without checking if it is filled.

Solution:
'l10n_mx_edi_customs_number' is an optional field, so we can add a verification to be sure it's filled before accessing it.

opw-2572849

closes odoo/enterprise#19114

X-original-commit: fb2776a634887ac4cf465752caf17a183c3683a1
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>
Signed-off-by: Djamel Touati <DjamelTouati@users.noreply.github.com>