[IMP] hr_appraisal_survey: show surveys from appraisal app
Surveys menuitem has been added in Appraisal Configuration.
Now it is possible to create surveys from appraisal app.
The surveys that are created from appraisal app (for appraisal purposes)
will only appear in the appraisal app.
Moreover, appraisal administrator has all CRUD rights on appraisal surveys.
Now, when we redirect to the survey results with the ‘Feedback Survey’ smart button,
we see in the header - appraisal info - for whom was the feedback requested, the date,
and the list of the requestors and requestees for the feedback.

task - 2528209

closes odoo/enterprise#19154

Related: odoo/upgrade#2574
Signed-off-by: Kevin Baptiste <kba@odoo.com>