[FIX] social: fix the link preview on the medias (Facebook and Twitter)
Bug
===
When we post from Odoo to Facebook or Twitter, the social media will
display a preview of the link (with the title, the description and the
logo of the HTML page). But when we post from Social, this preview
contains only "Redirecting..." for Facebook and nothing for Twitter.

The reason is that we use link tracker to register clicks on our link.

Here is what happen;
1. The social user write his post with an URL and click on post
2. The SQL transaction of the social user start
3. The link tracker is created
4. A HTTP request is made to the Facebook API
5. Facebook fetch the URL we give (link tracker) to build the preview
6. The SQL transaction of Facebook start
7. We try to search the code related to the link tracker, we do not find
   it because the SQL transaction of the social user is not yet
   committed
8. End of the SQL transaction of Facebook
9. So Facebook / Twitter consider that the link redirect to nowhere
   (Facebook show "Redirecting..." as title, Twitter nothing)
10. The Facebook API return the result to our HTTP request (point 4)
11. End of the transaction of the social user all changes are committed

So because Facebook / Twitter fetch our link tracker before we commit
the current transaction, the link tracker does not exist yet for them.

N.B. LinkedIn seems to fetch the link trackers after returning the
     response, so there's no issue for this media.

Task 2294034

closes odoo/enterprise#19162

X-original-commit: 927199986b5339cdd6c00d136a500867ee347d6b
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>