[FIX] *: convert createActionManager => createWebClient in tests
The legacy ActionManager has been deleted. Its features are handled by the couple
actionService + ActionContainer (Component). The top most level in which this becomes testable
is now the WebClient itself, which is the root component of the Odoo web application on `/web`.

This commit adapts all the tests that used an actionManager to make them use the webClient.
It also adapts other test suite to the change in how a test eexecution environment is created, set and cleaned.