[IMP] website_event_twitter_wall: revamp event template form
Adapt to community changes and improve the order in which the social_menu item
is displayed in both the event and event_type form.

Add the social_menu item to the event_type model to enable the user to show and
hide this menu item based on the template, thus aligning this submenu behaviour
with the other submenus, so that the user does not get confused.

Task-2541208

closes odoo/enterprise#19128

Com-pr: https://github.com/odoo/odoo/pull/71856
Upg-pr: https://github.com/odoo/upgrade/pull/2516
Related: odoo/odoo#71856
Related: odoo/upgrade#2516
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>