[FIX] l10n_mx_edi: Fix generated CFDI file for payments with multi-currency
The Mexican payment edi needs the link between the payment and how
much of every invoice was paid with it.  In multi-currency with a rate
that changes, that complicates matters.

This is why we check the reconciliation for the amounts in foreign currency
and the automatic exchange rate generated and  we also check the write-off,
that is an exchange rate difference (as Odoo reconciles in company currency (in this case),
it is possible that through the exchange rate difference, it should be considered
as paid already, so we need a manual exchange rate diff write-off).

That way we can construct an exchange rate at the moment of payment and the
amount of the invoice that corresponds to the payment currency.

closes odoo/enterprise#19683

Issue: 2448879
X-original-commit: e205067f326fec829c2eafa0ed90f4359a70fa6a
Signed-off-by: Josse Colpaert <jco@openerp.com>
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>