[FIX] web_studio: load wysiwyg with the right route
install website
Open a form view with description field (field_html)
Click edit
Enter studio with the systray

Before this commit, there was a crash because the urls of the wysiwyg asset loaded
the first time by the field html and the second time by the studio loader were different
to load the same asset. This was caused by the additional website_id that the public
route to load the wysiwyg (public_render_template) puts in the assets' urls.

After this commit, we always load the wysiwyg from the same route, so there is no crash.

closes odoo/enterprise#19139

Related: odoo/odoo#72454
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>