[FIX] stock_barcode: limit locations in pdf
Printing the barcode commands pdf will add all the picking types as well
as all the locations having a barcode. This can lead to a memory error
from wkhtmltopdf if there is too much data to handle.

This commit limits the number of locations and picking type sent to the
pdf generation.

The barcodes can be printed via the locations/picking types menus anyways
if needed.

closes odoo/enterprise#20181

Opw: 2564764
X-original-commit: ff582a8ed7f32381e66602ee04ea7a656a8aaae2
Related: odoo/odoo#74966
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>