[FIX] documents_spreadsheet: warning during test
"Can save a pivot in existing spreadsheet" would trigger a warning
about missing action "1".

It's unclear why but trying to provide the action via some sort of
`actions` key doesn't work, only intercepting `/web/action/load`
would. Strangely, when doing so the `write` call gets bumped to after
the end of the test function, and it does not seem to be a simple race
condition as simply adding delays (up to one second) before calling
"verifySteps" does not change the issue.

So... ignore it, just drop it from `verifySteps`. And update the rest
of the code a bit:

* Remove the `nextTick` calls, they don't seem to do anything.
* `assert.step` all methods called on `documents.document`, that seems
  like a good idea.

  Using a `switch` to handle these methods is a bit overkill as
  there's only one such case left (now that `write` and
  `get_spreadsheets_to_display` are both managed generically) but...