[FIX] web_studio: remove many2many_avatar_employee related test
With this commit we remove many2many_avatar_employee related test
statements as to test many2many_avatar_employee we require hr
modules, so if web_studio is installed alone then test case will
fail (for instance, on the runbot in single module builds).

Note that the scenario is still tested for the many2many_avatar_user
widget which is basically the same.

closes odoo/enterprise#19262

X-original-commit: a500e69d7e9b3f322416be6c5eb4581979480b81
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>