[REF] sign: make initDocumentToSign use existing session
the module depends on web.session already, meaning there is already a
promise to access the bound session. Rebinding the sesion can cause
issues especially with the new wowl infrastructure, as some services
will patch session_info and replace some fields with getters. Rebinding
the session will attempt to write on those non-writable fields, but in
this case there is no reason to do so.

Part of #19258