[IMP] stock_barcode: support GS1 barcodes
Add the support of GS1 nomenclature in `stock_barcode` module.
If the GS1 nomenclature is enabled, GS1 barcode can be scanned and its
data will be parsed and decomposed.
Some kind of data/records can be pass alongside other data in a single
barcode. These data are:
  - Product;
  - Quantity with an Unit of Measure (the UoM is ignored if the setting
    is disabled);
  - Tracking numbers;
  - Package.

The location will be recognized but will be treated separatly.

The GS1 standard specification aren't fully implemented as all data type
aren't usable in Odoo and this commit is only the first step for other
changes about GS1 implementation. See module `barcodes_gs1_nomenclature`
data rules to find currently supported rules.
So, some incomplete mandatory combinations or forbidden combinations are
possible in the barcode application and will be processed as possible.

task-2352413