[IMP] mail_enterprise, *: use explicit export instead of default
* = approvals, documents, sign, voip, web_studio

Exporting directly on the line of the class or variable definition is less lines
of code and less repetition (and risk or mistake).

Exporting with a name instead of default allows to catch typos more easily when
importing and ensures the same name is used for consistency (and ease of grep).

closes odoo/enterprise#19224

Related: odoo/odoo#72597
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>