[FIX] l10n_mx_edi: standardizes MX invoices columns' classes
In 14.0 when creating a Mexican CFDI, signing it, and then printing it, certain columns cannot span on multiple lines, and are thus forcing other columns to span on way too many lines to preserve horizontal space.

With this commit, the classes of the columns that cannot span are updated to match those of standard columns, resulting in a balanced layout.

opw-2524425

closes odoo/enterprise#19362

X-original-commit: 0b8f5ba009c15eb6afe9d051a90838420b483b34
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: pvh-odoo <SwagSamaSempai@users.noreply.github.com>