[FIX] web_grid: allows toggle in timesheets cells for all localizations
In 14.0 when toggling between different cell values in a timesheet grid
which is configured to work with days instead of hours, the cycling
(0 -> 0.5 -> 1) would not work in languages with decimal separator
different than point (eg. french), because of mix usage between native
format/parse JS function and odoo functions.

With this commit, we only use odoo functions and the the cycling values
(0 -> 0.5 -> 1) are not changed depending on the localization, which
means the dot is never deleted or replaced with a comma.

With the fix, the added test fails on two assertions:

    0,5 is the next value since 0,0 was the closest value in the range
    0,5 becomes 1,0 as it is the next value in the range

opw-2494212

closes odoo/enterprise#19385

X-original-commit: bd99dcdb6653b412090bce3b99702165f601a264
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: pvh-odoo <SwagSamaSempai@users.noreply.github.com>