[IMP] planning,project(_holidays): link an employee time off with tasks/slot
Purpose of the task is to display the warning stating that an employee
is on holiday at the time the task is planned would reinforce the feature.
In addition, the fact that a given cell is grey in the Gantt view can
easily be overlooked if a task was already planned for this user
before he took that day off.

So in this commit,
 - Added the module to integrate the tasks with time off and display the
warning on task when the user is on leave.
 - also added the filter to find the task on which the user is on leave
 - also move the conflict warning (and decoration) from Field Service to Project
  and added a 'tasks in conflict' filter on tasks.
 - Add warning msg for non-valiated time off and also include those
time off in filter too.
 - also move the common code to hr_holidays_gantt.

closes odoo/enterprise#19743

Taskid: 2453321
Related: odoo/upgrade#2803
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>